requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-admin-honeypot==1.1.0
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.4.0
django-seed==0.2.2
et-xmlfile==1.0.1
Faker==4.9.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==2.10
isort==5.5.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pylint==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
python-dotenv==0.19.2
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
reportlab==3.6.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.0
tablib==2.0.0
text-unidecode==1.3
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==1.2.0
xlwt==1.3.0
